requirements.txt
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-WTF==0.14.3
WTForms==2.3.3
mysqlclient==2.1.0
Werkzeug==1.0.1
SQLAlchemy==1.4.36
Jinja2==3.0.3
python-dotenv==0.19.2
email-validator==1.1.3
